requirements.txt
python-decouple==3.8
requests==2.31.0
python-dotenv==1.0.0
ccxt==4.1.13
pandas==2.1.4
numpy==1.26.2
loguru==0.7.2
matplotlib==3.8.2
seaborn==0.13.0 